+++ messages/ru-RU/app/modules/robots.php
@@ -3,6 +3,79 @@
 return [
     'Robots.txt' => "Robots.txt",
     'Generating and edit the `robots.txt` file' => "Генерация и редактирование файла `robots.txt`",
+
+    'ID' => 'ИД',
+    'Robot' => 'Робот',
+    'User-agent' => 'User-agent',
+    'Rule' => 'Правило',
+    'Rule path or value' => 'Путь правила или значение',
+    'Mode' => 'Режим',
+    'Directive' => 'Директива',
+    'Status' => "Статус",
+    'Description' => "Описание",
+    'Created at' => 'Создано',
+    'Updated at' => 'Обновлено',
+    'Created' => "Создано",
+    'Updated' => "Обновлено",
+    'Created by' => "Автор создания",
+    'Updated by' => "Автор изменений",
+
+    'Actions' => "Действия",
+    'Re-generate' => "Сгенерировать",
+    'View robots.txt' => "Просмотр robots.txt",
+    'Add new rule' => "Добавить правило",
+    'Update rule' => "Редиктирование правила",
+
+    'All robots' => "Все роботы",
+
+    'All statuses' => "Все статусы",
+    'Disabled' => "Автор изменений",
+    'Active' => "Автор изменений",
+
+    'All modes' => "Все режимы",
+    'Allow' => "Allow",
+    'Disallow' => "Disallow",
+    'Clean-Param' => "Clean-Param",
+    'Host' => "Host",
+    'Crawl-delay' => "Crawl-delay",
+    'Sitemap' => "Sitemap",
+
+    'Robots.txt by path `{path}` is not exists.' => "Robots.txt по пути `{path}` не существует.",
+    'Robots.txt by path `{path}` is not writable.' => "Robots.txt по пути `{path}` недоступен для записи.",
+
+    'Robots.txt has been successfully regenerated!' => "Robots.txt был успешно сгенерирован!",
+    'An error occurred while saving `robots.txt` file.' => "Произошла ошибка при сохранении файла `robots.txt`.",
+    'Robots.txt by path `{path}` is not exists or not writable.' => "Robots.txt по пути `{path}` не существует или недоступен для записи.",
+
+    'Robots.txt rule has been successfully added!' => 'Правило Robots.txt успешно добавлено!',
+    'An error occurred while add the rule.' => 'Произошла ошибка при добавлении правила.',
+
+    'Robots.txt rule has been successfully updated!' => 'Правило Robots.txt успешно обновлено!',
+    'An error occurred while updating the rule.' => 'Произошла ошибка при обновлении правила.',
+
+    'The requested rule does not exist.' => 'Запрашиваемое правило не существует.',
+
+    'OK! Rule successfully deleted.' => 'OK! Правило было успешно удалено!',
+    'An error occurred while deleting the rule.' => 'Произошла ошибка при удалении правила.',
+
+    'Close' => 'Закрыть',
+    '&larr; Back to list' => '&larr; Назад к списку',
+    'Save' => 'Сохранить',
+    'Add' => 'Добавить',
+    'Edit' => 'Редактировать',
+    'Delete' => 'Удалить',
+
+    'First page' => 'Первая страница',
+    'Last page' => 'Последняя страница',
+    '&larr; Prev page' => '&larr; Предыдущая страница',
+    'Next page &rarr;' => 'Следующая страница &rarr;',
+
+    'The value of the field `{attribute}` must be a URL provided that `{mode}`.' => 'Значение поля `{attribute}` должно быть URL при условии `{mode}`.',
+    'The value of the field `{attribute}` must be an integer provided that `{mode}`.' => 'Значение поля `{attribute}` должно быть целым числом при условии `{mode}`.',
+    'The value of the field `{attribute}` can contain only Latin letters, numbers and spaces, as well as wildcards: -, _, =,?, &,%, /,., $, ~, * provided that `{mode} `.' => 'Значение поля `{attribute}` может содержать только латинские буквы, цыфры и пробел, а также подстановочные знаки: -, _, =, ?, &, %, /, ., $, ~, * при условии `{mode}`.',
+    'The value of the field `{attribute}` can contain only Latin letters and numbers, as well as wildcards: -, _, =,?, &,%, /,., $, ~, * provided that `{mode}`.' => 'Значение поля `{attribute}` может содержать только латинские буквы и цыфры, а также подстановочные знаки: -, _, =, ?, &, %, /, ., $, ~, * при условии `{mode}`.',
+
+
 ];
 
 ?>