+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
  
+## 0.*
+ * Added CRUD for manage rules
+ 
 ## 0.0.1 (2020-06-03)
  * Added composer package
  * Added bootstrap, base module and models